Make link checker scan md or html files
