+++ tools/chainlink/go.mod
@@ -2,4 +2,7 @@ module chainlink
 
 go 1.19
 
-require github.com/yuin/goldmark v1.5.4
+require (
+	github.com/yuin/goldmark v1.5.4
+	golang.org/x/net v0.9.0
+)
