+++ tools/chainlink/types.go
@@ -13,9 +13,32 @@ import (
 	"github.com/yuin/goldmark/parser"
 )
 
+var (
+	// flags
+	hostname    = ""
+	scheme      = ""
+	httpMethod  = ""
+	ignoreFile  = ""
+	resultsFile = ""
+	checkAll    = false
+	extractMode = false
+	contentDir  = ""
+	fileType    = ""
+
+	// url rules
+	correctURLregex = &regexp.Regexp{}
+	ignores         = ignoreURLs{}
+
+	// url accumulators - should probably be fields of the results struct
+	checked   = linkAccumulator{Links: make(map[string]link), mu: &sync.Mutex{}}
+	unchecked = linkAccumulator{Links: make(map[string]link)}
+	ignored   = linkAccumulator{Links: make(map[string]link)}
+)
+
 type results struct {
 	Checked   []link `json:"checked"`
 	Unchecked []link `json:"unchecked"`
+	Ignored   []link `json:"ignored"`
 	json      []byte
 }
 
@@ -33,10 +56,17 @@ type pathWalker struct {
 	rawBytes []byte // the bytes read from a file
 }
 
-type nodeWalker struct {
+// for markdown
+type mdWalker struct {
 	path     string
 	rawBytes *[]byte // the bytes read from a file
 	parser   *parser.Parser
+}
+
+// for html
+type htmlWalker struct {
+	path     string
+	rawBytes *[]byte // the bytes read from a file
 }
 
 type ignoreURLs struct {
@@ -48,7 +78,7 @@ type ignoreURLs struct {
 type link struct {
 	URL     *url.URL               `json:"url"`
 	FullURL string                 `json:"fullurl"`
-	Status  int                    `json:"status,omitempty"`
+	Status  int                    `json:"status"`
 	Files   map[string]interface{} `json:"files"`
 	RawURL  string                 `json:"rawurl"`
 }
